Revert "quick fixup for id type"
This reverts commit d96f478043ed2caf5c7c06d58b0c04471f82f472.

+++ src/main/kotlin/net/xenotime/network/NetworkUtils.kt
@@ -4,11 +4,12 @@ import net.xenotime.network.game.proto.GamePacket
 import net.xenotime.network.game.proto.GameProto
 import net.xenotime.network.wormhole.proto.WormholePacket
 import net.xenotime.network.wormhole.proto.WormholeProto
+import java.util.*
 
 object NetworkUtils {
 
     object Game {
-        fun newPacket(id: Int): GamePacket? {
+        fun newPacket(id: Byte): GamePacket? {
             GameProto.values().filter {
                 it.id() == id
             }.let {
@@ -22,7 +23,7 @@ object NetworkUtils {
     }
 
     object Wormhole {
-        fun newPacket(id: Int): WormholePacket? {
+        fun newPacket(id: Byte): WormholePacket? {
             WormholeProto.values().filter {
                 it.id() == id
             }.let {
